Fix compilation errors in C# migration
- Fix for loop conditions and add missing short casts in AI_NPC.cs and praetorians.cs
- Extract repeated GetValue calls to local variables in FileIO.cs
- Replace inline lambdas with named local functions for better readability
- Remove readonly modifier from auxiliarBuffer in Protocol.cs
- Initialize posMadera variable in Trabajo.cs
- Add explicit short casts for bitwise operations in modSendData.cs
- Remove legacy Roslyn fix-err-number script (already applied)

All changes resolve C# compiler errors from VB.NET conversion.
Build now succeeds with 0 errors and 0 warnings.

+++ Legacy/Source/modSendData.cs
@@ -491,10 +491,10 @@ namespace Legacy
             {
                 tempIndex = Convert.ToInt16(ModAreas.ConnGroups[Map].UserEntrys[LoopC]);
 
-                TempInt = Declaraciones.UserList[tempIndex].AreasInfo.AreaReciveX & AreaX;
+                TempInt = (short)(Declaraciones.UserList[tempIndex].AreasInfo.AreaReciveX & AreaX);
                 if (TempInt != 0) // Esta en el area?
                 {
-                    TempInt = Declaraciones.UserList[tempIndex].AreasInfo.AreaReciveY & AreaY;
+                    TempInt = (short)(Declaraciones.UserList[tempIndex].AreasInfo.AreaReciveY & AreaY);
                     if (TempInt != 0)
                     {
                         if (tempIndex != UserIndex)
@@ -700,10 +700,10 @@ namespace Legacy
             {
                 tempIndex = Convert.ToInt16(ModAreas.ConnGroups[Map].UserEntrys[LoopC]);
 
-                TempInt = Declaraciones.UserList[tempIndex].AreasInfo.AreaReciveX & AreaX;
+                TempInt = (short)(Declaraciones.UserList[tempIndex].AreasInfo.AreaReciveX & AreaX);
                 if (TempInt != 0) // Esta en el area?
                 {
-                    TempInt = Declaraciones.UserList[tempIndex].AreasInfo.AreaReciveY & AreaY;
+                    TempInt = (short)(Declaraciones.UserList[tempIndex].AreasInfo.AreaReciveY & AreaY);
                     if (TempInt != 0)
                     {
                         if (Declaraciones.UserList[tempIndex].ConnIDValida)
@@ -737,10 +737,10 @@ namespace Legacy
             {
                 tempIndex = Convert.ToInt16(ModAreas.ConnGroups[Map].UserEntrys[LoopC]);
 
-                TempInt = Declaraciones.UserList[tempIndex].AreasInfo.AreaReciveX & AreaX;
+                TempInt = (short)(Declaraciones.UserList[tempIndex].AreasInfo.AreaReciveX & AreaX);
                 if (TempInt != 0) // Esta en el area?
                 {
-                    TempInt = Declaraciones.UserList[tempIndex].AreasInfo.AreaReciveY & AreaY;
+                    TempInt = (short)(Declaraciones.UserList[tempIndex].AreasInfo.AreaReciveY & AreaY);
                     if (TempInt != 0)
                     {
                         if (Declaraciones.UserList[tempIndex].ConnIDValida)
